delete file and write pdf on fly

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 exports.getAddProducts = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -140,6 +141,7 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             if (updatedImg) {
+                fileHelper.deleteFile(product.imgUrl);
                 product.imgUrl = updatedImg.path;
             }
             product.description = updatedDescription;
@@ -156,12 +158,19 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const productId = req.body.productId;
-    Product.deleteOne({ _id: productId, userId: req.user._id })
-        .then(() => {
-            res.redirect('/admin/products-list');
-        }).catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
-        });
+
+    Product.findById(productId).then( product => {
+        if (!product) {
+            return next(new Error('Product not found!'));
+        }
+        fileHelper.deleteFile(product.imgUrl);
+        return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
+    .then(() => {
+        res.redirect('/admin/products-list');
+    }).catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
